backend_fastapi/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.14.0
    # via
    #   -r requirements.in
    #   alembic-autogen-check
    #   alembic-postgresql-enum
alembic-autogen-check==1.1.1
    # via -r requirements.in
alembic-postgresql-enum==1.4.0
    # via -r requirements.in
amqp==5.3.1
    # via kombu
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via starlette
async-timeout==5.0.1
    # via redis
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
beautifulsoup4==4.12.3
    # via nbconvert
billiard==4.2.1
    # via celery
bleach==6.2.0
    # via nbconvert
boto3==1.35.73
    # via -r requirements.in
botocore==1.35.77
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.0
    # via google-auth
celery==5.4.0
    # via
    #   -r requirements.in
    #   flower
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   alembic-autogen-check
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
defusedxml==0.7.1
    # via nbconvert
deprecated==1.2.15
    # via limits
exceptiongroup==1.2.2
    # via anyio
fastapi==0.115.5
    # via -r requirements.in
fastapi-pagination==0.12.32
    # via -r requirements.in
fastjsonschema==2.21.1
    # via nbformat
flower==2.0.1
    # via -r requirements.in
google-api-core==2.24.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.154.0
    # via -r requirements.in
google-auth==2.36.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via
    #   -r requirements.in
    #   google-api-python-client
google-auth-oauthlib==1.2.1
    # via -r requirements.in
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via -r requirements.in
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.66.0
    # via google-api-core
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via uvicorn
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
humanize==4.11.0
    # via flower
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.4
    # via nbconvert
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.23.0
    # via
    #   -r requirements.in
    #   nbformat
jsonschema-specifications==2024.10.1
    # via
    #   -r requirements.in
    #   jsonschema
jupyter-client==8.6.3
    # via nbclient
jupyter-core==5.7.2
    # via
    #   jupyter-client
    #   nbclient
    #   nbconvert
    #   nbformat
jupyterlab-pygments==0.3.0
    # via nbconvert
kombu==5.4.2
    # via celery
language-tool-python==2.8.1
    # via -r requirements.in
limits==3.14.1
    # via slowapi
mako==1.3.8
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   nbconvert
mistune==3.0.2
    # via nbconvert
nbclient==0.10.1
    # via nbconvert
nbconvert==7.16.4
    # via -r requirements.in
nbformat==5.10.4
    # via
    #   -r requirements.in
    #   nbclient
    #   nbconvert
numpy==2.2.0
    # via pandas
oauthlib==3.2.2
    # via requests-oauthlib
packaging==24.2
    # via
    #   limits
    #   nbconvert
pandas==2.2.3
    # via -r requirements.in
pandocfilters==1.5.1
    # via nbconvert
platformdirs==4.3.6
    # via jupyter-core
prometheus-client==0.21.1
    # via flower
prompt-toolkit==3.0.48
    # via click-repl
proto-plus==1.25.0
    # via google-api-core
protobuf==5.29.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==6.1.0
    # via -r requirements.in
psycopg2==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.3
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-pagination
pydantic-core==2.27.1
    # via pydantic
pyjwt==2.3.0
    # via -r requirements.in
pygments==2.18.0
    # via nbconvert
pyparsing==3.2.0
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
    #   jupyter-client
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
python-multipart==0.0.19
    # via -r requirements.in
pytz==2024.2
    # via
    #   flower
    #   pandas
pyzmq==26.2.0
    # via jupyter-client
redis==5.2.0
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   language-tool-python
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
s3transfer==0.10.4
    # via boto3
six==1.17.0
    # via python-dateutil
slowapi==0.1.9
    # via -r requirements.in
sniffio==1.3.1
    # via anyio
soupsieve==2.6
    # via beautifulsoup4
sqlalchemy==2.0.36
    # via
    #   -r requirements.in
    #   alembic
    #   alembic-postgresql-enum
starlette==0.41.3
    # via fastapi
tinycss2==1.4.0
    # via nbconvert
tornado==6.4.2
    # via
    #   flower
    #   jupyter-client
tqdm==4.67.1
    # via language-tool-python
traitlets==5.14.3
    # via
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   fastapi-pagination
    #   limits
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   uvicorn
tzdata==2024.2
    # via
    #   celery
    #   kombu
    #   pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.3
    # via
    #   botocore
    #   requests
uvicorn==0.32.1
    # via -r requirements.in
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
wheel==0.45.1
    # via language-tool-python
wrapt==1.17.0
gitpython
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
